+++ gazebo_geometries_plugin/src/geometries_plugin.cpp
@@ -66,7 +66,7 @@ void GeometriesPlugin::initKinovaDimensions()
     Vector3 base, shoulder, arm, arm_half_1, arm_half_2, forearm, wrist, wrist_spherical_1, wrist_spherical_2, hand_3finger;
 
     // values are obtained by measuring mesh files in kinova_description package
-    link_finger_1.x = 0.13, link_finger_1.y = 0.025, link_finger_1.z = 0.025;
+    link_finger_1.x = 0.132, link_finger_1.y = 0.025, link_finger_1.z = 0.025;
     link_finger_2 = link_finger_3 = link_finger_1;
     
     base.x = 0.083, base.y = 0.083, base.z = 0.16;